+++ Editor/Editors/PropertyDrawers/YarnNodeAttributeDrawer.cs
@@ -5,6 +5,7 @@ Yarn Spinner is licensed to you under the terms found in the file LICENSE.md.
 using System.Collections.Generic;
 using UnityEditor;
 using UnityEngine;
+using Yarn.Unity.Attributes;
 
 namespace Yarn.Unity
 {