+++ Runtime/Utility/DialogueReference.cs
@@ -3,6 +3,7 @@ Yarn Spinner is licensed to you under the terms found in the file LICENSE.md.
 */
 
 using System;
+using Yarn.Unity.Attributes;
 
 namespace Yarn.Unity
 {